Added reward-penalty system
Player gets a score for each brick hit.

The score multiplier is a function of how many bricks they have hit with the current ball to reward them for their performance.

Multiplier increases logarithmically to avoid saving excessively high scores.

The score multiplier resets each time the ball is lost as a penalty.